--- a/core/trino-main/src/main/java/io/trino/dispatcher/LocalDispatchQueryFactory.java
+++ b/core/trino-main/src/main/java/io/trino/dispatcher/LocalDispatchQueryFactory.java
@@ -16,6 +16,7 @@ package io.trino.dispatcher;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.ListeningExecutorService;
 import io.airlift.log.Logger;
+import io.trino.FeaturesConfig;
 import io.trino.Session;
 import io.trino.event.QueryMonitor;
 import io.trino.execution.ClusterSizeMonitor;
@@ -63,6 +64,7 @@ public class LocalDispatchQueryFactory
     private final Map<Class<? extends Statement>, QueryExecutionFactory<?>> executionFactories;
     private final WarningCollectorFactory warningCollectorFactory;
     private final ListeningExecutorService executor;
+    private final boolean faultTolerantExecutionExchangeEncryptionEnabled;
 
     @Inject
     public LocalDispatchQueryFactory(
@@ -75,7 +77,8 @@ public class LocalDispatchQueryFactory
             Map<Class<? extends Statement>, QueryExecutionFactory<?>> executionFactories,
             WarningCollectorFactory warningCollectorFactory,
             ClusterSizeMonitor clusterSizeMonitor,
-            DispatchExecutor dispatchExecutor)
+            DispatchExecutor dispatchExecutor,
+            FeaturesConfig featuresConfig)
     {
         this.queryManager = requireNonNull(queryManager, "queryManager is null");
         this.transactionManager = requireNonNull(transactionManager, "transactionManager is null");
@@ -87,6 +90,7 @@ public class LocalDispatchQueryFactory
         this.warningCollectorFactory = requireNonNull(warningCollectorFactory, "warningCollectorFactory is null");
         this.clusterSizeMonitor = requireNonNull(clusterSizeMonitor, "clusterSizeMonitor is null");
         this.executor = dispatchExecutor.getExecutor();
+        this.faultTolerantExecutionExchangeEncryptionEnabled = requireNonNull(featuresConfig, "featuresConfig is null").isFaultTolerantExecutionExchangeEncryptionEnabled();
     }
 
     @Override
@@ -112,7 +116,8 @@ public class LocalDispatchQueryFactory
                 executor,
                 metadata,
                 warningCollector,
-                getQueryType(preparedQuery.getStatement()));
+                getQueryType(preparedQuery.getStatement()),
+                faultTolerantExecutionExchangeEncryptionEnabled);
 
         // It is important that `queryCreatedEvent` is called here. Moving it past the `executor.submit` below
         // can result in delivering query-created event after query analysis has already started.