--- a/core/trino-main/src/main/java/io/trino/execution/buffer/PagesSerdeFactory.java
+++ b/core/trino-main/src/main/java/io/trino/execution/buffer/PagesSerdeFactory.java
@@ -13,18 +13,12 @@
  */
 package io.trino.execution.buffer;
 
-import io.airlift.compress.lz4.Lz4Compressor;
-import io.airlift.compress.lz4.Lz4Decompressor;
-import io.airlift.slice.Slice;
 import io.trino.spi.block.BlockEncodingSerde;
-import io.trino.spiller.AesSpillCipher;
-import io.trino.spiller.SpillCipher;
 
-import javax.crypto.spec.SecretKeySpec;
+import javax.crypto.SecretKey;
 
 import java.util.Optional;
 
-import static com.google.common.base.Verify.verify;
 import static java.util.Objects.requireNonNull;
 
 public class PagesSerdeFactory
@@ -38,25 +32,8 @@ public class PagesSerdeFactory
         this.compressionEnabled = compressionEnabled;
     }
 
-    public PagesSerde createPagesSerde(Optional<Slice> encryptionKey)
+    public PagesSerde createPagesSerde(Optional<SecretKey> encryptionKey)
     {
-        return createPagesSerdeInternal(encryptionKey.map(key -> {
-            verify(key.hasByteArray(), "key is expected to be based on a byte array");
-            return new AesSpillCipher(new SecretKeySpec(key.byteArray(), key.byteArrayOffset(), key.length(), "AES"));
-        }));
-    }
-
-    public PagesSerde createPagesSerdeForSpill(Optional<SpillCipher> spillCipher)
-    {
-        return createPagesSerdeInternal(spillCipher);
-    }
-
-    private PagesSerde createPagesSerdeInternal(Optional<SpillCipher> spillCipher)
-    {
-        if (compressionEnabled) {
-            return new PagesSerde(blockEncodingSerde, Optional.of(new Lz4Compressor()), Optional.of(new Lz4Decompressor()), spillCipher);
-        }
-
-        return new PagesSerde(blockEncodingSerde, Optional.empty(), Optional.empty(), spillCipher);
+        return new PagesSerde(blockEncodingSerde, compressionEnabled, encryptionKey);
     }
 }