--- a/core/trino-main/src/main/java/io/trino/operator/output/PartitionedOutputOperator.java
+++ b/core/trino-main/src/main/java/io/trino/operator/output/PartitionedOutputOperator.java
@@ -227,7 +227,8 @@ public class PartitionedOutputOperator
                 sourceTypes,
                 maxMemory,
                 operatorContext,
-                positionsAppenderFactory);
+                positionsAppenderFactory,
+                operatorContext.getSession().getExchangeEncryptionKey());
 
         operatorContext.setInfoSupplier(this.partitionFunction.getOperatorInfoSupplier());
         this.memoryContext = operatorContext.newLocalUserMemoryContext(PartitionedOutputOperator.class.getSimpleName());