--- a/core/trino-main/src/main/java/io/trino/operator/output/TaskOutputOperator.java
+++ b/core/trino-main/src/main/java/io/trino/operator/output/TaskOutputOperator.java
@@ -27,6 +27,7 @@ import io.trino.operator.OutputFactory;
 import io.trino.spi.Page;
 import io.trino.spi.type.Type;
 import io.trino.sql.planner.plan.PlanNodeId;
+import io.trino.util.Ciphers;
 
 import java.util.List;
 import java.util.function.Function;
@@ -104,7 +105,7 @@ public class TaskOutputOperator
         this.operatorContext = requireNonNull(operatorContext, "operatorContext is null");
         this.outputBuffer = requireNonNull(outputBuffer, "outputBuffer is null");
         this.pagePreprocessor = requireNonNull(pagePreprocessor, "pagePreprocessor is null");
-        this.serde = serdeFactory.createPagesSerde(operatorContext.getSession().getExchangeEncryptionKey());
+        this.serde = serdeFactory.createPagesSerde(operatorContext.getSession().getExchangeEncryptionKey().map(Ciphers::deserializeAesEncryptionKey));
     }
 
     @Override
@@ -162,10 +163,8 @@ public class TaskOutputOperator
     {
         List<Page> split = splitPage(page, DEFAULT_MAX_PAGE_SIZE_IN_BYTES);
         ImmutableList.Builder<Slice> builder = ImmutableList.builderWithExpectedSize(split.size());
-        try (PagesSerde.PagesSerdeContext context = serde.newContext()) {
-            for (Page p : split) {
-                builder.add(serde.serialize(context, p));
-            }
+        for (Page p : split) {
+            builder.add(serde.serialize(p));
         }
         return builder.build();
     }