@losipiuk 2022-11-09
One more question. Can the page size grow now compared to what it used to be (due to encryption). If so we may run into the problem coming from the default limit on HTTP response size imposed by airlift http server. I think it is 16MB by default.