@arhimondr 2022-11-09
Benchmarks

```
+----------------------+-----------------------+----------------------+-----------------------+----------+-----------+
| base_cpu_time_millis | base_wall_time_millis | test_cpu_time_millis | test_wall_time_millis | cpu_diff | wall_diff |
+----------------------+-----------------------+----------------------+-----------------------+----------+-----------+
|          18184316267 |              77292978 |          18135469806 |              76886882 |  0.99731 |   0.99475 |
+----------------------+-----------------------+----------------------+-----------------------+----------+-----------+

+-------------------------------+----------------------+-----------------------+----------------------+-----------------------+----------+-----------+
| suite                         | base_cpu_time_millis | base_wall_time_millis | test_cpu_time_millis | test_wall_time_millis | cpu_diff | wall_diff |
+-------------------------------+----------------------+-----------------------+----------------------+-----------------------+----------+-----------+
| tpcds_sf10000_partitioned     |           2151027639 |              10666623 |           2068382944 |               9833001 |  0.96158 |   0.92185 |
| tpcds_sf10000_partitioned_etl |          12575752138 |              45907876 |          12553147695 |              45802726 |  0.99820 |   0.99771 |
| tpcds_sf100_partitioned       |             25285611 |               1298684 |             24391767 |               1155394 |  0.96465 |   0.88967 |
| tpcds_sf100_partitioned_etl   |            112666635 |               5309658 |            117064847 |               5243526 |  1.03904 |   0.98754 |
| tpch_sf10000_bucketed         |            833789592 |               3341612 |            811495718 |               3408108 |  0.97326 |   1.01990 |
| tpch_sf10000_bucketed_etl     |           2459310174 |               9295363 |           2533861412 |               9857522 |  1.03031 |   1.06048 |
| tpch_sf100_bucketed           |              6506475 |                168296 |              5445624 |                152070 |  0.83695 |   0.90359 |
| tpch_sf100_bucketed_etl       |             19978003 |               1304866 |             21679799 |               1434535 |  1.08518 |   1.09937 |
+-------------------------------+----------------------+-----------------------+----------------------+-----------------------+----------+-----------+


```

Details

https://gist.github.com/arhimondr/b200e1d76752a2dc0f1063ab13863dcf