@arhimondr 2022-11-09
> Can the page size grow now compared to what it used to be (due to encryption)

It can by 256 bit (plaintext IV) + 248 bit (max padding) + 256 (encrypted IV for verification) + 32 bit (encrypted length) with the total of 99 bytes per block (64kb by default) what should in theory result into ~0.15% increase in page size. So in theory there's a tiny chance of pushing a query over the limit. 

> If so we may run into the problem coming from the default limit on HTTP response size imposed by airlift http server

We don't enable encryption for pipelined execution. However in fault tolerant execution there's also a limit on max page size. But in fault tolerant execution we also enable encryption that makes the limits even more "obscure".