@arhimondr 2022-11-10
Micro benchmarks:

Before:

```
Benchmark                        (compressed)  (encrypted)  (randomSeed)   Mode  Cnt     Score    Error  Units
BenchmarkPagesSerde.deserialize          true         true          1000  thrpt   10  1194.138 ±  7.442  ops/s
BenchmarkPagesSerde.deserialize          true        false          1000  thrpt   10  2281.985 ±  8.634  ops/s
BenchmarkPagesSerde.deserialize         false         true          1000  thrpt   10  1416.509 ±  2.928  ops/s
BenchmarkPagesSerde.deserialize         false        false          1000  thrpt   10  5891.232 ± 14.932  ops/s
BenchmarkPagesSerde.serialize            true         true          1000  thrpt   10   312.647 ±  0.800  ops/s
BenchmarkPagesSerde.serialize            true        false          1000  thrpt   10   601.053 ±  1.522  ops/s
BenchmarkPagesSerde.serialize           false         true          1000  thrpt   10   451.260 ±  0.996  ops/s
BenchmarkPagesSerde.serialize           false        false          1000  thrpt   10  3446.756 ±  6.949  ops/s
```

After:

```
Benchmark                        (compressed)  (encrypted)  (randomSeed)   Mode  Cnt     Score    Error  Units
BenchmarkPagesSerde.deserialize          true         true          1000  thrpt   10  1732.440 ±  4.044  ops/s
BenchmarkPagesSerde.deserialize          true        false          1000  thrpt   10  3059.334 ±  5.412  ops/s
BenchmarkPagesSerde.deserialize         false         true          1000  thrpt   10  2093.146 ±  3.505  ops/s
BenchmarkPagesSerde.deserialize         false        false          1000  thrpt   10  5906.148 ± 13.923  ops/s
BenchmarkPagesSerde.serialize            true         true          1000  thrpt   10   329.150 ±  1.269  ops/s
BenchmarkPagesSerde.serialize            true        false          1000  thrpt   10   617.507 ±  1.808  ops/s
BenchmarkPagesSerde.serialize           false         true          1000  thrpt   10   493.182 ±  0.687  ops/s
BenchmarkPagesSerde.serialize           false        false          1000  thrpt   10  3816.939 ±  7.984  ops/s
```